added other tools; update docs;

+++ install.sh
@@ -2,20 +2,21 @@
 set -o nounset    # error when referencing undefined variable
 set -o errexit    # exit when command fails
 
+echo 'Copying .dotfiles'
 ln -s $(pwd) ~/.dotfiles
 
 # Script to install my files
-echo 'Installing .vimrc'
-
-echo 'Copying to ~/.vimrc'
-cp ~/.vimrc ~/.vimrc_backup
+echo 'Copying .vimrc'
+[ -f ~/.vimrc ] && cp ~/.vimrc ~/.vimrc_backup
+[ -f ~/.config/nvim/init.vim ] && cp ~/.config/nvim/init.vim ~/.config/nvim/init.vim_backup
 ln -s `pwd`/.vimrc ~/.vimrc
-
-echo 'Copying to ~/.config/nvim/init.vimrc'
-cp ~/.config/nvim/init.vim ~/.config/nvim/init.vim_backup
 ln -s `pwd`/.vimrc ~/.config/nvim/init.vim
 
+echo 'Copying .tmux.conf'
+[ -f ~/.tmux.conf ] && cp ~/.tmux.conf ~/.tmux.conf
+ln -s `pwd`/.tmux.conf ~/.tmux.conf
 
+echo 'Installing coc-vim'
 # for neovim
 mkdir -p ~/.local/share/nvim/site/pack/coc/start
 cd ~/.local/share/nvim/site/pack/coc/start
